Infrastructure to add debugging hooks to DDlog operators.
The first step towards implementing a DDlog debugger.  The idea is to
instrument the DDlog program with debugging hooks in a way that does not
require program rewriting (and hence recompilation) and does not
increase the memory footprint of the program being debugged.  To this
end, we will (1) instrument each operator with a debugging callback
that, when enabled, notifies the debugger about each invocation of the
operator and its inputs, and (2) attach and `inspect()` operator to each
collection to track collection update.  The debugger can then store,
process, and present this raw data in a user-friendly way.

In this commit:

- Added a post-processing callback to `XFormArrangement::Antijoin`.  All
other operators already take user-defined functions, which can be used
to introduce debugging hooks.  Antijoin was the only one that didn't.

- Changed the compiler to produce two implementations of each
user-defined function, with and without debugging hooks.  Added a flag
(`__debug__`) to the generated Rust code that determines at runtime,
specifically when the dataflow is being assembled, whether debugging
hooks should be enabled.