Introduce #[multiway] attribute for rules.
We introduce the `#[multiway]` attribute that indicates that a rule must be
evaluated as a multiway join. It does not yet do anything meaningful
except that it prevents common prefix elimination for multiway rules. In
addition, this commit improves parsing of attributes in general.  In
particular, it allows associating multiple attributes with specification
entities and allows attributes without explicit value (in which case the
value is assumed to be `true`).  Finally, we add the position field to
the `RHS` type to improve error reporting in the future.