Refactor validation code.
Most validation code used to be in `Validate.hs`.  A better design is to
keep validation methods together with other code that manipulates the
same entity.  We move expression validation code to `Expr.hs`, rule
validation code to `Rule.hs`, and type validation code to `Type.hs`.

This commit also makes a couple of improvements to actual validation
logic:

- `exprFreeVars` method returns the set of free variables in an
  expression, i.e., variables used in the expression, but declared
  outside of it.  This enables more accurate implementation of
  `exprIsConst`, `exprIsInjective`, `ruleRHSTermVars`, and
  `arrangeInputs` methods.

- Added code to validate that a rule can be evaluated as a multiway join
  (currently only possible for a very restricted set of rules)

Finally, this commit adds two methods to re-arrange terms in the
body of the rule, which will be needed to re-shuffle rules for
multiway joins:
- `ruleMoveCondsUp` - optimizes a rule by moving all conditions in the
  body to the earliest possible position.
- `ruleOptimizeJoins` - optimizes the rule by rearranging join
  operations to maximize the number of variables shared by subsequent
  joins.