Store small values inline.
Dynamic dispatch, as implemented in the previous commits, introduces a
large number of small heap allocations, causing significant CPU and
memory overhead.  This is particularly wasteful for small values, as
each value, even if it's only a few bytes long, requires 32 bytes
of metadata (pointer to Arc + pointer to vtable + 8-byte strong
reference counter + 8-byte weak reference counter).

We eliminate this overhead for small values <=8 bytes by storing them
inline inside `DDValue` (where `Arc<value_type>`) would normally be
stored.  Unfortunately, this means that we can no longer rely on Rust's
dynamic dispatch mechanism, which does not support such optimizations.
We mock up our own dynamic dispatch mechanism instead.  Instead of
using trait objects, we define our own datatype that consists of a
pointer to data (or data stored inline) and a vtable, defined as a
struct with pointers.  At runtime we delegate operations on `DDValue`
to methods in the vtable.

Combined with previous optimizations, this improves the performance of
dynamic dispatch to be on a par with static dispatch on the redist_opt
benchmark, with memory footprint being even better than with static
dispatch (by ~15%).