Use servo_arc instead of std::sync::Arc.
The standard Arc type we use stores weak reference count in addition to
the strong one.  We don't use weak references, to this memory gets
wasted.

This commit replaces the standard Arc with a clone copied from the
servo project.  This seems to improve the memory footprint, but only
marginally.